requirements.txt
discord.py
python-dotenv
cryptography
bcrypt
